move key's function to keys.go
