+++ app/keepers/keepers.go
@@ -112,19 +112,6 @@ type AppKeepers struct {
 	memKeys map[string]*sdk.MemoryStoreKey
 }
 
-func (appKeepers *AppKeepers) GenerateKeys() {
-	// Define what keys will be used in the cosmos-sdk key/value store.
-	// Cosmos-SDK modules each have a "key" that allows the application to reference what they've stored on the chain.
-	// Keys are in keys.go to kep app.go as brief as possible.
-	appKeepers.keys = sdk.NewKVStoreKeys(KVStoreKeys()...)
-
-	// Define transient store keys
-	appKeepers.tkeys = sdk.NewTransientStoreKeys(paramstypes.TStoreKey)
-
-	// MemKeys are for information that is stored only in RAM.
-	appKeepers.memKeys = sdk.NewMemoryStoreKeys(capabilitytypes.MemStoreKey)
-}
-
 func (appKeepers *AppKeepers) InitNormalKeepers(
 	appCodec codec.Codec,
 	bApp *baseapp.BaseApp,
@@ -524,42 +511,4 @@ func KVStoreKeys() []string {
 		bech32ibctypes.StoreKey,
 		wasm.StoreKey,
 	}
-}
-
-func (appKeepers *AppKeepers) GetSubspace(moduleName string) paramstypes.Subspace {
-	subspace, _ := appKeepers.ParamsKeeper.GetSubspace(moduleName)
-	return subspace
-}
-
-func (appKeepers *AppKeepers) GetKVStoreKey() map[string]*sdk.KVStoreKey {
-	return appKeepers.keys
-}
-
-func (appKeepers *AppKeepers) GetTransientStoreKey() map[string]*sdk.TransientStoreKey {
-	return appKeepers.tkeys
-}
-
-func (appKeepers *AppKeepers) GetMemoryStoreKey() map[string]*sdk.MemoryStoreKey {
-	return appKeepers.memKeys
-}
-
-// GetKey returns the KVStoreKey for the provided store key.
-//
-// NOTE: This is solely to be used for testing purposes.
-func (appKeepers *AppKeepers) GetKey(storeKey string) *sdk.KVStoreKey {
-	return appKeepers.keys[storeKey]
-}
-
-// GetTKey returns the TransientStoreKey for the provided store key.
-//
-// NOTE: This is solely to be used for testing purposes.
-func (appKeepers *AppKeepers) GetTKey(storeKey string) *sdk.TransientStoreKey {
-	return appKeepers.tkeys[storeKey]
-}
-
-// GetMemKey returns the MemStoreKey for the provided mem key.
-//
-// NOTE: This is solely used for testing purposes.
-func (appKeepers *AppKeepers) GetMemKey(storeKey string) *sdk.MemoryStoreKey {
-	return appKeepers.memKeys[storeKey]
 }
